# packages/proxy/src/openapi.yaml
openapi: 3.1.0
info:
  title: HAAI API
  description: |
    Management REST API and OpenAI-compatible inference endpoints for HAAI.

    **Management API** (`/api/v1/*`) requires admin authentication via session cookie or Bearer token.

    **Inference API** (`/v1/*`) requires an API key in the `Authorization: Bearer` header.
  version: 0.0.1
servers:
  - url: /
    description: Current host
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: haai_session
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: http
      scheme: bearer
      description: Client API key (haai-sk-…)
...
paths:
...
  /api/v1/backends:
    get:
      tags: [Backends]
      summary: List backends
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Backend list
    post:
      tags: [Backends]
      summary: Create backend
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, hostName, provider, baseUrl]
              properties:
                name: { type: string }
                displayName: { type: string }
                hostName: { type: string }
                provider: { type: string, example: lmstudio }
                baseUrl: { type: string, format: uri }
                keyMode: { type: string, enum: [passthrough, abstraction] }
                apiKey: { type: string }
                enabled: { type: boolean }
                weight: { type: integer }
      responses:
        "201":
          description: Backend created
  /api/v1/backends/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Backends]
      summary: Get backend
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Backend details
        "404":
          description: Not found
    patch:
      tags: [Backends]
      summary: Update backend
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated backend
    delete:
      tags: [Backends]
      summary: Delete backend
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "204":
          description: Deleted
  /api/v1/backends/{id}/test:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Backends]
      summary: Test backend connectivity
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Test result with latency and discovered models
...
  /api/v1/hooks:
    get:
      tags: [Hooks]
      summary: List hooks
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Hook list
    post:
      tags: [Hooks]
      summary: Create hook
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "201":
          description: Hook created
  /api/v1/hooks/{id}/test:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Hooks]
      summary: Test hook
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Hook test result
...